update filter lib
