+++ src/lib/dataFilter.test.js
@@ -4,6 +4,24 @@
 //   filterByLevelAndDomain,
 //   orderConfences,
 // } from './dataFilter.lib';
+import { getTags } from './dataFilter.lib';
+
+const conferences = [
+  {
+    title: 'conf1',
+    tags: ['foo', 'bar'],
+  },
+  {
+    title: 'conf2',
+    tags: ['foo', 'bar', 'test'],
+  },
+];
+
+describe('database.lib', () => {
+  it('should get tags', async () => {
+    expect((await getTags(conferences)).length).toBe(3);
+  });
+});
 
 it('should filter conferences by domain', () => {
   //   expect(filterByDomain(data, 'tech').length).toBe(19);
