+++ src/lib/database.js
@@ -1,4 +1,3 @@
-import _ from 'lodash';
 import { db } from '../firebase';
 
 const dbRef = (suffix = '') => db.ref(`2017${suffix}`);
@@ -10,7 +9,3 @@ export const writeStore = (state) => {
 export const readStore = async id => (await db.ref('/users').get(id)).val();
 
 export const getConferences = async () => (await dbRef('/conferences').once('value')).val();
-
-export const getTags = async conferences =>
-  // const conferences = await getConferences();
-  _.uniq(_.flatten(_.map(conferences, item => item.tags)));
